Property components

+++ src/service/UserService.ts
@@ -1,17 +1,20 @@
 import { IUserService } from '../interface/IUserService';
 import { IUser } from '../interface/IUser';
 import { IUserRepository } from '../interface/IUserRepository';
-import { UserRepository } from '../repositories/UserRepository';
 
 export class UserService implements IUserService {
   private userRepository: IUserRepository;
 
-  constructor() {
-    this.userRepository = new UserRepository();
+  constructor(userRepository: IUserRepository) {
+    this.userRepository = userRepository;
   }
 
   async createUser(user: IUser): Promise<IUser> {
-    return await this.userRepository.createUser(user);
+    const userExists = await this.userRepository.getUserByEmail(user.email);
+    if(userExists) {
+      throw new Error(`User with ID ${user.id} already exists`);
+    }
+    return await this.userRepository.addUser(user);
   }
 
   async getUserById(id: number): Promise<IUser | null> {
@@ -23,10 +26,18 @@ export class UserService implements IUserService {
   }
 
   async updateUser(id: number, userData: Partial<IUser>): Promise<IUser | null> {
+    const userExists = await this.userRepository.getUserById(id);
+    if(!userExists){
+      throw new Error(`User with ID ${id} does not exist`);
+    }
     return await this.userRepository.updateUser(id, userData);
   }
 
   async deleteUser(id: number): Promise<boolean> {
+    const userExists = await this.userRepository.getUserById(id);
+    if(!userExists){
+      throw new Error(`User with ID ${id} does not exist`);
+    }
     return await this.userRepository.deleteUser(id);
   }
 }
